requirements.txt
fastapi
uvicorn
moviepy==1.0.3
jinja2
boto3
python-multipart
snowflake-connector-python